exampleの結果の表示部分を修正
std::cout<<を使うとき、数値はintで渡さないとちゃんと数字で表示できなかったので修正

+++ example/example.cpp
@@ -20,9 +20,9 @@ int main(int argc, char **argv)
 
     Mseq mseq(N);
 
-    for (int i = 0; i < (1 << N) -1; i++)
+    for (int i = 0; i < (1 << N) - 1; i++)
     {
-        std::cout << mseq.update() << " ";
+        std::cout << (int)mseq.update() << " ";
     }
     std::cout << std::endl;
